hotfix for db

+++ app/src/main/java/com/ryan/heroestopbuilds/HeroDatabase.java
@@ -40,6 +40,8 @@ public class HeroDatabase extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        if (oldVersion >= newVersion)
+            return;
         db.execSQL("DROP TABLE IF EXISTS " + TABLE_HEROES);
         onCreate(db);
     }
